+++ internal/db/db.go
@@ -62,42 +62,3 @@ func Store(shortID, originalURL string) error {
 	_, err := DB.Exec("INSERT INTO urls (short_id, original_url) VALUES ($1, $2)", shortID, originalURL)
 	return err
 }
-
-// GetData - возвращает данные из базы данных в виде map[string]string,
-// где ключ - short_id, значение - original_url.
-func GetData() (data map[string]string, err error) {
-	if !IsConnected() {
-		return nil, errors.New("GetData. No connection to DB")
-	}
-
-	rows, err := DB.Queryx("SELECT short_id, original_url FROM urls")
-	if err != nil {
-		return nil, err
-	}
-	defer rows.Close()
-
-	if rows.Err() != nil {
-		return nil, rows.Err()
-	}
-
-	data = make(map[string]string)
-
-	for rows.Next() {
-		var shortID, originalURL string
-		err = rows.Scan(&shortID, &originalURL)
-		if err != nil {
-			log.Warn().Err(err).Msg("GetData Cannot scan row")
-			continue
-		}
-		data[shortID] = originalURL
-	}
-
-	return data, err
-}
-
-// // LoadData - загружает данные из map[string]string, где ключ - short_id, значение - original_url, в базу данных.
-// func LoadData(data map[string]string) {
-// 	for shortID, originalURL := range data {
-// 		Store(shortID, originalURL)
-// 	}
-// }
