+++ internal/filestorage/filestorage.go
@@ -1,3 +1,11 @@
+// Description: Файловое хранилище для хранения записей в формате JSON.
+// Пример содержимого файла хранилища:
+// ```json
+// {"uuid":"1","short_url":"4rSPg8ap","original_url":"http://yandex.ru"}
+// {"uuid":"2","short_url":"edVPg3ks","original_url":"http://ya.ru"}
+// {"uuid":"3","short_url":"dG56Hqxm","original_url":"http://practicum.yandex.ru"}
+// ```
+
 package filestorage
 
 import (
@@ -6,18 +14,10 @@ import (
 	"path/filepath"
 
 	"github.com/google/uuid"
-	"github.com/rs/zerolog/log"
 	"github.com/vadim-ivlev/url-shortener/internal/config"
-	"github.com/vadim-ivlev/url-shortener/internal/storage"
 )
 
 // fileStorageRecord - структура для хранения записи в файловом хранилище.
-// Пример содержимого файла хранилища:
-// ```json
-// {"uuid":"1","short_url":"4rSPg8ap","original_url":"http://yandex.ru"}
-// {"uuid":"2","short_url":"edVPg3ks","original_url":"http://ya.ru"}
-// {"uuid":"3","short_url":"dG56Hqxm","original_url":"http://practicum.yandex.ru"}
-// ```
 type fileStorageRecord struct {
 	UUID        string `json:"uuid"`
 	ShortURL    string `json:"short_url"`
@@ -39,11 +39,10 @@ func createDirIfNotExists(filePath string) error {
 
 // Store - сохраняет данные в файловое хранилище.
 // Параметры:
-// - path - путь к файлу хранилища.
 // - shortURL - укороченный URL.
 // - originalURL - оригинальный URL.
 // Возвращает ошибку, если запись не удалась.
-func Store(path, shortURL, originalURL string) error {
+func Store(shortURL, originalURL string) error {
 	// Генерируем новый UUID
 	uuid, err := uuid.NewV7()
 	if err != nil {
@@ -63,12 +62,12 @@ func Store(path, shortURL, originalURL string) error {
 	}
 
 	// Создаем директорию для файла хранилища, если ее нет
-	if err := createDirIfNotExists(path); err != nil {
+	if err := createDirIfNotExists(config.Params.FileStoragePath); err != nil {
 		return err
 	}
 
 	// Открываем файл для записи (добавляем в конец файла) или создаем новый
-	file, err := os.OpenFile(path, os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0644)
+	file, err := os.OpenFile(config.Params.FileStoragePath, os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0644)
 	if err != nil {
 		return err
 	}
@@ -79,44 +78,4 @@ func Store(path, shortURL, originalURL string) error {
 		return err
 	}
 	return nil
-}
-
-// LoadData - загружает данные из файлового хранилища в storage.
-// Параметры:
-// - path - путь к файлу хранилища.
-// Возвращает срез записей и ошибку, если чтение не удалось.
-func LoadData(path string) ([]fileStorageRecord, error) {
-	// Открываем файл для чтения
-	file, err := os.OpenFile(path, os.O_RDONLY|os.O_CREATE, 0644)
-	if err != nil {
-		return nil, err
-	}
-	defer file.Close()
-
-	// Читаем все записи из файла
-	records := make([]fileStorageRecord, 0)
-	decoder := json.NewDecoder(file)
-	for {
-		var record fileStorageRecord
-		if err := decoder.Decode(&record); err != nil {
-			break
-		}
-		records = append(records, record)
-
-		// Извлекаем shortID из record.ShortURL
-		shortID := record.ShortURL[len(config.Params.BaseURL)+1:]
-		// Добавляем запись в карту хранилища
-		storage.Set(shortID, record.OriginalURL)
-	}
-	return records, nil
-}
-
-// LoadDataAndLog - загружает данные из файлового хранилища и выводит сообщение в лог.
-func LoadDataAndLog(path string) {
-	_, err := LoadData(path)
-	if err != nil {
-		log.Warn().Err(err).Msg("Filestorage not found. Probably this is the first launch.")
-	} else {
-		log.Info().Msg("Filestorage loaded")
-	}
 }
